Move footer, phase, cookies and feedback to core
